Add an empty contributor file.
This is a workaround to make the CI action work as expected.
The CLA action failed in this PR:
https://github.com/cowprotocol/flash-loan-wrapper-solver/pull/1

The error in the action log is:

```
Run contributor-assistant/github-action@v2.2.1
CLA Assistant GitHub Action bot has started the process
Error: Could not retrieve repository contents: Not Found. Status: 404
Error: Cannot read properties of undefined (reading 'data')
```

See for example:
https://github.com/cowprotocol/flash-loan-wrapper-solver/actions/runs/13071434597/job/36473778174

This workaround was suggested in:
https://github.com/contributor-assistant/github-action/issues/155#issuecomment-2190310649