Add library to store data to transient storage (#1)
I introduce a library contract that can be imported to write a byte
array from `calldata` to `transient` storage.

This will be needed to write the data for the call to `settle` expected
by the solver once the flash-loan contract calls back to this wrapper
contract. Many implementations of flash loans allow to send some data in
the callback, but we don't want to trust the flash-loan contract to send
us back the right data to call: then, since we lose the call context
when invoking the flash-loan contract, the cheapest reliable way to pass
data across calls is transient storage.

The implementation is done by having the library provide two functions:
one to write a bytes array to transient storage and the other to read
from it.

This is going to be by far the most low-level and technical PR of this
project to review. The main reason for that is that transient storage
isn't yet well supported by the Solidity compiler, with `transient`
variables having only been introduced with Solidity v0.8.28 and only for
value type (i.e., no `bytes` nor fixed-size array). I tried to convey as
many of the low level details as possible as inline comments. Efficient
copy from calldata to transient storage is still likely to require
assembly even with better Solidity support.

### How to test

New unit tests.

### Note on gas consumption

The new tests can be used to estimate the cost of writing and reading
through transient storage.
This is done by replacing `sequentialByteArrayOfSize` with a function
that just allocates an empty array of the given length.
Running the tests with this change we get the following expected
overhead:

<details><summary>Gas overhead</summary>

| Bytes in a call to `settle` | gas overhead | gas overhead per byte | 
|-----------------------------|--------------|-----------------------|
| 0                           | 17114        | n/a                   |
| 2                           | 17606        | 8803                  |
| 32                          | 17717        | 553.65                |
| 33                          | 18216        | 552.00                |
| 42784                       | 748761       | 17.50                 |
| 42785                       | 749449       | 17.52                 |
| 2560001                     | 596891022    | 233.16                |

The reason for the overhead increase per byte in the last test is the
quadratic growth of storing data in memory compared to the linear growth
of using transient storage and is likely to never affect us in a real
transaction.

</details>

A simple settlement (trade against private liquidity) is about 1.2kB,
which would be `33754` gas in this test (28.13 gas/B). The test with
42kB shows the result of using reasonable calldata size for a very
complex settlement.

---------

Co-authored-by: Martin Magnus <martin.beckmann@protonmail.com>