Add function to clear transient storage array (#3)
Code to clear the transient storage array. For gas efficiency, the
actual transient storage is not cleared, however it's inaccessible after
clearing. This function allows the transient storage array to be used as
some kind of reentrancy guard (see #2).

Note that I also sneaked in a bug fix: a test was marked as `private`
and not `external`, meaning that it wasn't actually being run. As a
future improvement process, we should check with a linter that functions
of the form `test_*` are either `public` or `external`.

### How to test

New fuzz tests.