Add router and borrower interfaces (#10)
This PR introduces the components of the flash loan support for CoW
Protocol in the form of interfaces (or minimal libraries) and documents
them extensively.

The main component is `IFlashLoanRouter`: this contract manages all
flash-loan requests and is eventually responsible for executing the
settlement.

`IBorrower` is an abstraction around specific flash-loan providers. Each
provider has slightly different syntax for the loan request, as well as
for the function that is called back. The borrower contract abstracts
this away for the router.

There are a few accessory interfaces as well to describe what contracts
we work with.

### How a settlement works

The entry point of a transaction is `flashLoanAndSettle()` in
`IFlashLoanRouter`.

You can see the call sequence of an example settlement in the following
diagram. The settlement happens after two flash loans have been
requested: one first from Aave and then another one from Maker (each of
which is expected to have a dedicated borrower contract that implements
`IBorrower`.

Each arrow is an EVM call, _not_ a call scope. For example, the
settlement happens in the eight nested call (at least, depending on the
internals of the lenders). Contracts marked with an asterisk `*` are
outside of the control of this repository.

```mermaid
sequenceDiagram
  FlashLoanRouter ->> Aave borrower: flashLoanAndCallBack()
  Aave borrower ->> Aave lender contract*: Aave-specific flash loan request
  Aave lender contract* ->> Aave borrower: Aave-specific flash loan callback
  Aave borrower ->> FlashLoanRouter: borrowerCallBack()
  FlashLoanRouter ->> Maker borrower: flashLoanAndCallBack()
  Maker borrower ->> Maker lender contract*: Maker-specific flash loan request
  Maker lender contract* ->> Maker borrower: Maker-specific flash loan callback
  Maker borrower ->> FlashLoanRouter: borrowerCallBack()
  FlashLoanRouter ->> CoW Settlement*: settle()
```

## How to test

Review the documentation and confirm that the design is consistent with
what was discussed in #9 (here, `FlashLoanSolverWrapper` has been
renamed to `Borrower`).

---------

Co-authored-by: Felix Leupold <felixleupold90@gmail.com>