Add LoansWithSettlement library (#11)
Extends the `LoansWithSettlement` library with some simple functions to
interact with the `LoansWithSettlement.Data` struct.

The current implementation is gas inefficient (in line with the one
discussed in [this
comment](https://github.com/cowprotocol/flash-loan-wrapper-solver/pull/9#discussion_r1973961796)),
so it will eventually be replaced by an implementation that does custom
memory management through assembly (a preview of which can be found
[here](https://github.com/cowprotocol/flash-loan-wrapper-solver/pull/9/files#diff-4b49b1d8e1069995667808cf2bacd83677a8ddf3c06add649c827d73f8c2d370)).
We introduce first the simpler version for ease of reviewing:
eventually, seeing the diff between the new assembly code with the
current base should be easier.

The test are more advanced than needed for the current version of the
library and are intended to eventually confirm that the switch with
custom memory management has worked well. They will be reused as they
are with minimal adjustments.

We also remove the existing transient-storage-based library as we'll be
using the `LoansWithSettlement` to help forwarding call data in each
call instead of using transient storage as discussed in the related [gas
experiment
PR](https://github.com/cowprotocol/flash-loan-wrapper-solver/pull/6).

### How to test

Unit tests. Confirm in the review that the unit tests are robust in case
of faulty low-level memory management.

---------

Co-authored-by: Felix Leupold <felixleupold90@gmail.com>