Add borrower for Aave & e2e test (#16)
Introduces another borrower for Aave support together with the related
e2e test.
All tests are very similar to the ones for ERC-3156/Maker and it's
mostly consistent with the work already done in #4 and #8.

Some interesting discussion points:
1. Aave has a flash-loan fee. I imagine we can try to ask them to waive
it at some point, but for now it's good to see some change in behavior
in the test.
2. The Aave lender natively supports multiple flash loans at once,
unlike ERC-3156. The current Borrower format does *not* take advantage
of this as it only supports one loan at a time per borrower. It wouldn't
be too hard to extend the router to send a list of loans to the
borrower, however I don't plan to implement this before the audit.

### How to test

New tests.

---------

Co-authored-by: Felix Leupold <felixleupold90@gmail.com>