Drop "wrapper" in naming convention (#17)
At first, borrowers were independent contracts and were called "solver
wrappers." There are some leftovers of this naming convention,
especially in the e2e tests. This PR makes the naming consistent across
the repo.

### How to test

Search for "wrapper" in the repo and only find an unrelated entry in a
vendored library.