Gas efficiency with manual memory management (#20)
Introduce the gas optimizations from manual memory management from the
draft PRs.

There are two commits: very easy optimizations (first commit) and
complex optimizations (second one).

The gas improvements in the second commit are particularly noticeable
for large settlements. They come from taking control of memory
management at a low level and trying as much as possible to avoid
copying memory.

### Gas improvements

Here is a comparison between the first and second commit, so that we can
judge the impact of the extra complexity.

|Gas profile | Maker before | Maker after | Aave before | Aave after |
Aave&Maker before | Aave&Maker after
|-|-:|-:|-:|-:|-:|-:|
| Average | 319,629 | 308,556 (-11,073) | 393,114 | 378,521 (-14,593) |
621,089 | 586,643 (-34,446) |
| Max | 736,062 | 597,403 (-138,659) | 1,053,357 | 913,788 (-139,569) |
1,527,042 | 1,203,392 (-323,650) |
| Median | 299,084 | 291,718 (-7,366) | 349,441 | 339,496 (-9,945) |
566,270 | 542,460 (-23,810) |
| Min | 283,047 | 278,191 (-4,856) | 314,688 | 307,894 (-6,794) |
523,416 | 506,791 (-16,625) |

It's also interesting to compare with the benchmarks from the draft PR
#9: the final gas cost has gone down for the average case of ~8kgas
(Maker), ~10kgas (Aave), and ~20kgas (combined) thanks to the
improvements in the code.

### How to test

New unit tests and existing test infrastructure. Note that in #11 we
already wrote the test to handle incorrect memory encoding.