Linting with Solhint (#23)
I've run Solhint on the current codebase and it found a few issues
(mostly unused imports and missing visibility markers).
This PR addresses them (except those in vendored contracts).

It would be nice to add this to the CI process but it would require
adding quite a bit of JS infrastructure, and this caused a few issues in
the past. For now, we'll only run linting manually.

<details><summary>.solhint.json for src/</summary>

```json
{
  "extends": "solhint:recommended",
  "rules": {
    "func-visibility": ["warn", { "ignoreConstructors": true }],
    "compiler-version": "off",
    "gas-custom-errors": "off",
    "immutable-vars-naming": "off",
    "avoid-low-level-calls": "off",
    "no-inline-assembly": "off"
  }
}
```
</details>

<details><summary>.solhint.json for test/</summary>

```json
{
  "extends": "solhint:recommended",
  "rules": {
    "func-visibility": ["warn", { "ignoreConstructors": true }],
    "compiler-version": "off",
    "gas-custom-errors": "off",
    "immutable-vars-naming": "off",
    "avoid-low-level-calls": "off",
    "func-name-mixedcase": "off",
    "one-contract-per-file": "off",
    "no-inline-assembly": "off",
    "reason-string": "off",
    "no-empty-blocks": "off",
    "payable-fallback": "off"
  }
}
```
</details>

### How to test

Existing tests.