Add Settlement event (#29)
Address a point raised in the Ackee audit about the fact that CoW
Protocol's `Settlement` event reports the address of the router instead
of the original solver. This PR adds a new event that emits the address
of the actual solver in the context of the router for disambiguation.

Internal discussion on back-end needs
[here](https://cowservices.slack.com/archives/C0375NV72SC/p1742899831311889).

### How to test

New test

### Gas analysis

#### Aave


| Gas profile | Before | This PR |
|-|-:|-:|
| Average | 378,499 | 379,646 (+1,147) |
| Max | 913,766 | 914,913 (+1,147) |
| Median | 339,474 | 340,621 (+1,147) |
| Min | 307,872 | 309,019 (+1,147) |

#### Aave then Maker

| Gas profile | Before | This PR |
|-|-:|-:|
| Average | 586,621 | 587,768 (+1,147) |
| Max | 1,203,370 | 1,204,517 (+1,147) |
| Median | 542,438 | 543,585 (+1,147) |
| Min | 506,769 | 507,916 (+1,147) |

#### Maker

| Gas profile | Before | This PR |
|-|-:|-:|
| Average | 308,538 | 309,456 (+918) |
| Max | 597,381 | 598,528 (+1,147) |
| Median | 291,700 | 292,618 (+918) |
| Min | 278,174 | 279,091 (+917) |