Workaround for checking fmt on CI (#32)
We're currently ignoring the formatting check on CI because of a Foundry
issue.
With these changes, we ignore only the only file that are affected by
that issue.

The change only apply to the `ci` profile because I find it handy to
format the affected file by just removing the `transient` keywords,
format that file, and then put them back.

See how `fmt` configuration works
[here](https://book.getfoundry.sh/config/#standalone-sections).

### How to test

Check out CI logs and confirm `cargo fmt` is working as expected.