Add Ackee audit (#36)
(This is the same as PR #33 but issues with git history rewriting caused
it to be using an incorrect history. When reviewing, you want to make
extremely sure that the commit mentioned on the audit is on `main`.)

Adds the final audit report from Ackee Blockchain Security.
All found issues have been addressed in the latest commits.

### How to test

Check the PDF, see that commit
`f9c18679df0f5cda19611b05fca1c982786071eb` is on `main` as well as the
audited code.