Add Gnosis audit (#37)
Adds the second audit for the router contract, this time by Gnosis.

The audited commits are the same as for Ackee.

### How to test

Same as in #33: read the text and confirm that there are no issues and
the mentioned commits are on main
(f9c18679df0f5cda19611b05fca1c982786071eb and
930914fd8b73a9f9c6a628beada616110fb8ec35).