deploy erc3156 borrower + restore networks.json + deploy in ALL networks (#38)
This PR brings back the lost `networks.json` and make some changes in
the deployment run files, and the scripts to make sure next devs have an
easier time than I did.

Additionally deploys the `erc3156 borrower contract`.

## Idempotence
The scripts deploy some contracts using CREATE2.

When the contract doesn't exist, it will deploy it in the deterministic
address.

If someone deployed the contract already, the scripts would fail because
of the collision (before this PR):

![image](https://github.com/user-attachments/assets/4ee17782-2c94-481b-b4a8-adb6a16dec9f)

This PR makes a check if the contract is deployed or not. If not
deployed, then it will continue making the scripts indepmpotent. This
property makes it easier for the next person that wants to deploy a new
contract, only the contracts that haven't been deployed will be
deployed.

<img width="697" alt="image"
src="https://github.com/user-attachments/assets/8fe0fdb5-8e7e-448d-9adf-ca6eb60b8334"
/>


## Fix salt
The previous contracts were deployed using a different salt. We had to
recover it from the transaction params, and use it in order to generate
the expected addresses.


## Networks
~I deployed `mainnet` and `sepolia` happy to deploy for any other
network.~
~@MartinquaXD can you confirm if you need it in any other network?~

Deployed in: 
- `mainnet`
- `sepolia`
- `gnosis`
- `base`
- `arbitrum`

EDIT: 9/4/2025
Added
- Polygon
- Avalanche


## Manually fix some deployment run.json files
Someone deployed the contracts but didn't upload the run.json

This is quite inconvenient because this is an important context that
give us details about the creation transaction.

Furthermore, without this file we can't automatically generate the
`networks.json` expected by other teams, using
`dev/generate-networks-file.sh`

😅 Fixing this was quite some tedious work, because this had been
deployed to all supported chain for 2 of the 3 contracts, so It was a
lot of manual work to fix that (`cast` helped a lot though).

## Additional notes
I left some comments in the PR with hints on the code changes.

---------

Co-authored-by: Federico Giacon <58218759+fedgiac@users.noreply.github.com>