Test that deployment script returns expected addresses (#39)
We had a few issues in recovering deterministic addresses because the
nonce specified in this repo didn't match the one used when the
contracts were deployed.
To make sure this doesn't happen again and that we don't introduce
changes that lead to a changing deployment address, we test to make sure
the contracts are deployed at the current addresses.

I wrote the code on top of #38 so that it also included the
ERC3156Borrower from there.

This PR is loosely based on #35 and it supersedes it.

### How to test

See that CI works.

---------

Co-authored-by: Anxo Rodriguez <anxolin@gmail.com>
Co-authored-by: Anxo Rodriguez <anxo@cow.fi>