feat: remove the preHook flag (#48)
### GIST
Backend can't provide flashloan assets to fulfill the verification of
the validity of the order.
I ended up removing the preHook check from the isValidSignature.

### THE FIX
instead of the order sending the funds to the user, now the order output
goes to the contract.
To avoid an attacker forcing the user to take the limit price without
the positive slippage, we are checking that the preHook was called.

If the attacker calls postHook() to force the limit price, they would
need to donate atokens to the helper contract.