Programmatically update networks.json (#52)
New networks were
[introduced](https://github.com/cowprotocol/flash-loan-router/pull/51)
but networks.json wasn't updated.
This PR introduces a script (copied from
https://github.com/cowdao-grants/cow-shed/commit/9e572403a38ef76607ec3146cc28fc2de8ea047b)
to account for manual additions in networks.json. This is where we get
the Lens deployment from.

The process used to be scripted but it was
[broken](https://github.com/cowprotocol/flash-loan-router/issues/46)
with Lens. I made it clear in the readme that the file should be
updated.

~I remember @anxolin has an updated version of the deployment script
that we can use to make this automated again but I can't find it. Please
let me know where I copy it from! (But this isn't a blocker for
merging.)~ (Added, thanks!)

I also pinned the Foundry version since standard formatting `forge fmt`
changed with the latest version.

### How to test

Make sure all chains in `broadcast/` are there and the pair address/tx
match. Verify that the script matches 1-to-1 with the one from the
linked commit.